double LL

+++ LL.cpp
@@ -25,7 +25,7 @@ class Node{
 Node* convertArr2LL(vector<int> &arr){
     Node* head =new Node(arr[0]);
     Node* mover=head;
-    for(int i=1;i<arr.size();i++)
+    for(size_t i=1;i<arr.size();i++)
     {
         Node* temp=new Node(arr[i]);
         mover->next=temp;
@@ -98,8 +98,9 @@ Node* insertPosition(Node*head ,int el , int k)
     int cnt=0;
     Node* temp= head;
     while(temp!=NULL)
-     {cnt++;
-     if(cnt==k-1)
+     {
+        cnt++;
+    if(cnt==k-1)
     {
         Node* x=new Node(el, temp->next);
         temp->next=x;
@@ -129,7 +130,7 @@ int main()
 
     // cout<<endl<<" Element is"<<searchinLL(head,val);
     //  head=removesTail(head);
-         head=insertPosition(head,100,3);
+    head=insertPosition(head,100,3);
 
      print(head);
